+++ src/utils/listUtils.ts
@@ -1,0 +1,3 @@
+const notEmpty = <T>(data: T[]) => data.length > 0;
+
+export { notEmpty };
