+++ webpack.config.js
@@ -1,0 +1,19 @@
+const path = require("path");
+
+module.exports = {
+  entry: "./src/weboodi",
+  output: {
+    path: path.resolve(__dirname, "dist"),
+    filename: "weboodi.js"
+  },
+  resolve: {
+    // Add `.ts` and `.tsx` as a resolvable extension.
+    extensions: [".ts", ".tsx", ".js"]
+  },
+  module: {
+    rules: [
+      // all files with a `.ts` or `.tsx` extension will be handled by `ts-loader`
+      { test: /\.tsx?$/, loader: "ts-loader" }
+    ]
+  }
+};
